+++ packages/ilmomasiina-backend/src/routes/ical/index.ts
@@ -2,6 +2,7 @@ import { FastifyReply, FastifyRequest } from "fastify";
 import { createEvents, DateArray } from "ics";
 import { Op } from "sequelize";
 
+import { getBranding } from "../../branding";
 import config, { eventDetailsUrl } from "../../config";
 import { Event } from "../../models/event";
 
@@ -38,9 +39,11 @@ export async function eventsAsICal() {
     ],
   });
 
+  const calendarName = (await getBranding()).icalCalendarName ?? config.icalCalendarName;
+
   const { error, value } = createEvents(
     events.map((event) => ({
-      calName: config.icalCalendarName,
+      calName: calendarName,
       uid: `${event.id}@${uidDomain}`,
       start: dateToArray(event.date!),
       startInputType: "utc",