+++ packages/ilmomasiina-frontend/src/routes/Login/index.tsx
@@ -7,10 +7,10 @@ import { useTranslation } from "react-i18next";
 import { useNavigate } from "react-router";
 
 import { errorDesc } from "@tietokilta/ilmomasiina-client";
-import branding from "../../branding";
 import FieldFormGroup from "../../components/FieldFormGroup";
 import type { TKey } from "../../i18n";
 import { loginToast } from "../../modules/auth";
+import { useEffectiveBranding } from "../../modules/branding";
 import useStore from "../../modules/store";
 import paths from "../../paths";
 import useEvent from "../../utils/useEvent";
@@ -31,6 +31,7 @@ const Login = () => {
   const { login } = useStore((state) => state.auth);
   const navigate = useNavigate();
   const { t } = useTranslation();
+  const branding = useEffectiveBranding();
 
   const onSubmit = useEvent(async (data: FormData) => {
     const { email, password } = data;